+++ DevotionTweaks/Devotion/DevotionTweaks.cs
@@ -13,6 +13,9 @@ using UE = UnityEngine;
 using LemurFusion.Devotion.Components;
 using RoR2.Projectile;
 using static RoR2.FriendlyFireManager;
+using HarmonyLib;
+using RoR2.Skills;
+using RoR2BepInExPack.GameAssetPaths;
 
 namespace LemurFusion.Devotion
 {
@@ -56,11 +59,38 @@ namespace LemurFusion.Devotion
 
             IL.RoR2.SceneDirector.PopulateScene += PopulateScene;
 
+            On.RoR2.AffixAurelioniteBehavior.OnServerDamageDealt += AffixAurelioniteBehavior_OnServerDamageDealt;
+            On.EntityStates.LemurianMonster.Bite.OnEnter += Bite_OnEnter;
+            On.EntityStates.LemurianBruiserMonster.Flamebreath.FireFlame += Flamebreath_FireFlame;
+
             if (PluginConfig.disableTeamCollision.Value)
             {
                 On.RoR2.BulletAttack.DefaultFilterCallbackImplementation += BulletAttack_DefaultFilterCallbackImplementation;
                 On.RoR2.Projectile.ProjectileController.IgnoreCollisionsWithOwner += ProjectileController_IgnoreCollisionsWithOwner;
             }
+        }
+
+        private static void Flamebreath_FireFlame(On.EntityStates.LemurianBruiserMonster.Flamebreath.orig_FireFlame orig, EntityStates.LemurianBruiserMonster.Flamebreath self, string muzzleString)
+        {
+            if (self.bulletAttack != null)
+                self.bulletAttack.damageType.damageSource = DamageSource.Secondary;
+
+            orig(self, muzzleString);
+        }
+
+        private static void Bite_OnEnter(On.EntityStates.LemurianMonster.Bite.orig_OnEnter orig, EntityStates.LemurianMonster.Bite self)
+        {
+            orig(self);
+
+            self.attack.damageType.damageSource = DamageSource.Secondary;
+        }
+
+        private static void AffixAurelioniteBehavior_OnServerDamageDealt(On.RoR2.AffixAurelioniteBehavior.orig_OnServerDamageDealt orig, AffixAurelioniteBehavior self, DamageReport damageReport)
+        {
+            if (Utils.IsDevoted(damageReport.attackerBody, out _) && !damageReport.damageInfo.damageType.IsDamageSourceSkillBased)
+                return;
+
+            orig(self, damageReport);
         }
 
         #region Misc
@@ -86,7 +116,7 @@ namespace LemurFusion.Devotion
 
             foreach (var tc in TeamComponent.GetTeamMembers(TeamIndex.Player))
             {
-                if (Utils.IsDevoted(tc.body) && tc.body.hurtBoxGroup && tc.gameObject != self.owner)
+                if (Utils.IsDevoted(tc.body, out _) && tc.body.hurtBoxGroup && tc.gameObject != self.owner)
                 {
                     var hurtBoxes = tc.body.hurtBoxGroup.hurtBoxes;
                     for (int i = 0; i < hurtBoxes.Length; i++)
@@ -102,19 +132,21 @@ namespace LemurFusion.Devotion
 
         private void LoadAssets()
         {
-            bodyPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/CU8/DevotedLemurianBody.prefab").WaitForCompletion();
+            bodyPrefab = Addressables.LoadAssetAsync<GameObject>(RoR2_CU8.DevotedLemurianBody_prefab).WaitForCompletion();
             var body = bodyPrefab.GetComponent<CharacterBody>();
             body.bodyFlags |= CharacterBody.BodyFlags.Devotion | CharacterBody.BodyFlags.ImmuneToLava | CharacterBody.BodyFlags.Mechanical;
             body.baseMaxHealth = 360f;
             body.levelMaxHealth = 11f;
             body.baseMoveSpeed = 7f;
 
-            bigBodyPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/CU8/DevotedLemurianBruiserBody.prefab").WaitForCompletion();
+            bigBodyPrefab = Addressables.LoadAssetAsync<GameObject>(RoR2_CU8.DevotedLemurianBruiserBody_prefab).WaitForCompletion();
             var bigBody = bigBodyPrefab.GetComponent<CharacterBody>();
             bigBody.bodyFlags |= CharacterBody.BodyFlags.Devotion | CharacterBody.BodyFlags.ImmuneToLava | CharacterBody.BodyFlags.Mechanical;
             bigBody.baseMaxHealth = 720f;
-            bigBody.levelMaxHealth = 22f;
+            bigBody.levelMaxHealth = 11f;
             bigBody.baseMoveSpeed = 10f;
+            bigBody.baseArmor = 0;
+            bigBody.levelArmor = 0.5f;
 
             var betterLemInvCtrlPrefab = LegacyResourcesAPI.Load<GameObject>("Prefabs/NetworkedObjects/DevotionMinionInventory");
             UE.Object.DestroyImmediate(betterLemInvCtrlPrefab.GetComponent<DevotionInventoryController>());
@@ -125,16 +157,28 @@ namespace LemurFusion.Devotion
                 BetterInventoryController.activationSoundEventDef = asset;
             });
 
-            masterPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/CU8/LemurianEgg/DevotedLemurianMaster.prefab").WaitForCompletion();
+            masterPrefab = Addressables.LoadAssetAsync<GameObject>(RoR2_CU8_LemurianEgg.DevotedLemurianMaster_prefab).WaitForCompletion();
             UE.Object.DestroyImmediate(masterPrefab.GetComponent<DevotedLemurianController>());
             masterPrefab.AddComponent<BetterLemurController>();
             masterPrefab.AddComponent<Inventory>();
             masterPrefab.GetComponent<CharacterMaster>().bodyPrefab = bodyPrefab;
+
             bruiserMasterPrefab = masterPrefab.InstantiateClone(devotedBruiserMasterName, true);
             bruiserMasterPrefab.GetComponent<CharacterMaster>().bodyPrefab = bigBodyPrefab;
 
-            var itemDef = Addressables.LoadAssetAsync<ItemDef>("RoR2/CU8/Harness/LemurianHarness.asset").WaitForCompletion();
+            Addressables.LoadAssetAsync<SkillDef>(RoR2_Base_LemurianBruiser.LemurianBruiserBodyPrimary_asset).Completed +=
+                (task) => task.Result.interruptPriority = EntityStates.InterruptPriority.Any;
+
+            Addressables.LoadAssetAsync<SkillDef>(RoR2_Base_LemurianBruiser.LemurianBruiserBodySecondary_asset).Completed +=
+                (task) => task.Result.interruptPriority = EntityStates.InterruptPriority.Any;
+
+            var itemDef = Addressables.LoadAssetAsync<ItemDef>(RoR2_CU8_Harness.LemurianHarness_asset).WaitForCompletion();
             itemDef.tags = [.. itemDef.tags.Concat([ItemTag.BrotherBlacklist, ItemTag.CannotSteal]).Distinct()];
+
+            var fireball = Addressables.LoadAssetAsync<GameObject>(RoR2_Base_Lemurian.Fireball_prefab).WaitForCompletion();
+            fireball.GetComponent<ProjectileDamage>().damageType.damageSource = DamageSource.Primary;
+            var fireball2 = Addressables.LoadAssetAsync<GameObject>(RoR2_Base_LemurianBruiser.LemurianBigFireball_prefab).WaitForCompletion();
+            fireball.GetComponent<ProjectileDamage>().damageType.damageSource = DamageSource.Primary;
         }
         #endregion
 
@@ -182,7 +226,7 @@ namespace LemurFusion.Devotion
                         if (minionOwnership && minionOwnership.TryGetComponent<BetterLemurController>(out var lemCtrl) && lemCtrl.LemurianInventory)
                         {
                             lemCount++;
-                            var fusionCount = lemCtrl.LemurianInventory.GetItemCount(CU8Content.Items.LemurianHarness);
+                            var fusionCount = lemCtrl.FusionCount;
                             if (fusionCount < lowestCount)
                             {
                                 lowestCount = fusionCount;
@@ -204,21 +248,45 @@ namespace LemurFusion.Devotion
             var c = new ILCursor(il);
 
             if (c.TryGotoNext(MoveType.After,
-                    i => i.MatchCall<RunArtifactManager>("get_instance"),
+                    i => i.MatchCallOrCallvirt(AccessTools.PropertyGetter(typeof(RunArtifactManager), nameof(RunArtifactManager.instance))),
                     i => i.MatchLdsfld("RoR2.CU8Content/Artifacts", "Devotion"),
-                    i => i.MatchCallvirt<RunArtifactManager>(nameof(RunArtifactManager.IsArtifactEnabled))
+                    i => i.MatchCallOrCallvirt<RunArtifactManager>(nameof(RunArtifactManager.IsArtifactEnabled))
+                ))
+            {
+                c.Emit(OpCodes.Pop);
+                c.Emit(OpCodes.Ldc_I4_0);
+            }
+            else
+            {
+                LemurFusionPlugin.LogError("IL Hook failed for SceneDirector_PopulateScene 1");
+            }
+
+            if (c.TryGotoNext(i => i.MatchNewobj(typeof(DirectorPlacementRule))) &&
+                c.TryGotoNext(i => i.MatchNewobj(typeof(DirectorSpawnRequest))) &&
+                c.TryGotoPrev(MoveType.After,
+                    x => x.MatchCallOrCallvirt<DirectorCard>(nameof(DirectorCard.GetSpawnCard))
                 ))
             {
                 c.Emit(OpCodes.Ldarg_0);
-                c.EmitDelegate<Func<bool, SceneDirector, bool>>((isArtifactEnabled, self) =>
+                c.EmitDelegate<Func<SpawnCard, SceneDirector, SpawnCard>>((originalCard, self) =>
                 {
-                    return (PluginConfig.permaDevotion.Value || isArtifactEnabled) &&
-                            self.rng.RangeInt(0, 100) < PluginConfig.eggSpawnChance.Value;
+                    if (!DevotionInventoryController.isDevotionEnable)
+                        return originalCard;
+
+                    if (!(originalCard is InteractableSpawnCard originalIsc && originalIsc.skipSpawnWhenDevotionArtifactEnabled))
+                        return originalCard;
+
+                    if (self.rng.RangeInt(0, 100) < Math.Clamp(PluginConfig.eggSpawnChance.Value, 0, 100))
+                    {
+                        return Addressables.LoadAssetAsync<SpawnCard>(RoR2_CU8_LemurianEgg.iscLemurianEgg_asset).WaitForCompletion();
+                    }
+
+                    return originalCard;
                 });
             }
             else
             {
-                LemurFusionPlugin.LogError("IL Hook failed for SceneDirector_PopulateScene");
+                LemurFusionPlugin.LogError("IL Hook failed for SceneDirector_PopulateScene 2");
             }
         }
         #endregion