+++ DevotionTweaks/Devotion/LemurControllerTweaks.cs
@@ -1,4 +1,5 @@
 ﻿using RoR2;
+using UnityEngine;
 
 namespace LemurFusion.Devotion
 {